some refactoring and clean-ups
