+++ plugins/org.eclipse.n4js.dts/src/org/eclipse/n4js/dts/astbuilders/DtsTypeRefBuilder.java
@@ -21,10 +21,7 @@ import static org.eclipse.n4js.dts.TypeScriptParser.RULE_unionTypeExpression;
 
 import java.util.Set;
 
-import org.eclipse.emf.common.util.URI;
-import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EReference;
-import org.eclipse.emf.ecore.InternalEObject;
 import org.eclipse.n4js.dts.DtsTokenStream;
 import org.eclipse.n4js.dts.TypeScriptParser.ColonSepTypeRefContext;
 import org.eclipse.n4js.dts.TypeScriptParser.ParameterizedTypeRefContext;
@@ -40,15 +37,7 @@ import org.eclipse.n4js.ts.typeRefs.TypeRefsPackage;
 import org.eclipse.n4js.ts.types.Type;
 import org.eclipse.n4js.ts.types.TypesFactory;
 import org.eclipse.n4js.ts.types.TypingStrategy;
-import org.eclipse.xtext.XtextFactory;
 import org.eclipse.xtext.linking.lazy.LazyLinkingResource;
-import org.eclipse.xtext.nodemodel.BidiTreeIterable;
-import org.eclipse.xtext.nodemodel.ICompositeNode;
-import org.eclipse.xtext.nodemodel.ILeafNode;
-import org.eclipse.xtext.nodemodel.INode;
-import org.eclipse.xtext.nodemodel.SyntaxErrorMessage;
-import org.eclipse.xtext.util.ITextRegion;
-import org.eclipse.xtext.util.ITextRegionWithLineInformation;
 
 /**
  * Builder to create {@link TypeReferenceNode} from parse tree elements
@@ -93,9 +82,7 @@ public class DtsTypeRefBuilder extends AbstractDtsSubBuilder<TypeRefContext, Typ
 
 		Type typeProxy = TypesFactory.eINSTANCE.createType();
 		EReference eRef = TypeRefsPackage.eINSTANCE.getParameterizedTypeRef_DeclaredType();
-		int fragmentNumber = resource.addLazyProxyInformation(pTypeRef, eRef, new PseudoLeafNode(text));
-		URI encodedLink = resource.getURI().appendFragment("|" + fragmentNumber);
-		((InternalEObject) typeProxy).eSetProxyURI(encodedLink);
+		ParserContextUtil.installProxy(resource, pTypeRef, eRef, typeProxy, text);
 		pTypeRef.setDeclaredType(typeProxy);
 
 		if (ctx.typeArguments() != null && ctx.typeArguments().typeArgumentList() != null) {
@@ -121,161 +108,5 @@ public class DtsTypeRefBuilder extends AbstractDtsSubBuilder<TypeRefContext, Typ
 		default:
 			return false;
 		}
-	}
-
-	static class PseudoLeafNode implements ILeafNode {
-		final String text;
-
-		PseudoLeafNode(String text) {
-			this.text = text;
-		}
-
-		@Override
-		public String getText() {
-			return text;
-		}
-
-		// Note that all methods below are irrelevant since they won't be used later on instances of this class
-
-		@Override
-		public ICompositeNode getParent() {
-			return null;
-		}
-
-		@Override
-		public boolean hasSiblings() {
-			return false;
-		}
-
-		@Override
-		public boolean hasPreviousSibling() {
-			return false;
-		}
-
-		@Override
-		public boolean hasNextSibling() {
-			return false;
-		}
-
-		@Override
-		public INode getPreviousSibling() {
-			return null;
-		}
-
-		@Override
-		public INode getNextSibling() {
-			return null;
-		}
-
-		@Override
-		public ICompositeNode getRootNode() {
-			return null;
-		}
-
-		@Override
-		public Iterable<ILeafNode> getLeafNodes() {
-			return null;
-		}
-
-		@Override
-		public int getTotalOffset() {
-			return 0;
-		}
-
-		@Override
-		public int getOffset() {
-			return 0;
-		}
-
-		@Override
-		public int getTotalLength() {
-			return 0;
-		}
-
-		@Override
-		public int getLength() {
-			return 0;
-		}
-
-		@Override
-		public int getTotalEndOffset() {
-			return 0;
-		}
-
-		@Override
-		public int getEndOffset() {
-			return 0;
-		}
-
-		@Override
-		public int getTotalStartLine() {
-			return 0;
-		}
-
-		@Override
-		public int getStartLine() {
-			return 0;
-		}
-
-		@Override
-		public int getTotalEndLine() {
-			return 0;
-		}
-
-		@Override
-		public int getEndLine() {
-			return 0;
-		}
-
-		@Override
-		public EObject getGrammarElement() {
-			return XtextFactory.eINSTANCE.createKeyword();
-		}
-
-		@Override
-		public EObject getSemanticElement() {
-			return null;
-		}
-
-		@Override
-		public boolean hasDirectSemanticElement() {
-			return false;
-		}
-
-		@Override
-		public SyntaxErrorMessage getSyntaxErrorMessage() {
-			return null;
-		}
-
-		@Override
-		public BidiTreeIterable<INode> getAsTreeIterable() {
-			return null;
-		}
-
-		@Override
-		public ITextRegion getTextRegion() {
-			return null;
-		}
-
-		@Override
-		public ITextRegion getTotalTextRegion() {
-			return null;
-		}
-
-		@Override
-		public ITextRegionWithLineInformation getTextRegionWithLineInformation() {
-			return null;
-		}
-
-		@Override
-		public ITextRegionWithLineInformation getTotalTextRegionWithLineInformation() {
-			return null;
-		}
-
-		@Override
-		public boolean isHidden() {
-			return false;
-		}
-
 	}
 }
